#202 BUILD: move to Typescript V7. Plan & Step 1
