+++ package.json
@@ -20,6 +20,7 @@
     "typescript": "^6.0.2"
   },
   "scripts": {
+    "postinstall": "python scripts/patch-tsup-baseurl.py",
     "nonreg": "python scripts/run-nonreg.py",
     "nonreg:unit": "python scripts/run-nonreg.py --tier unit --run-all",
     "nonreg:default": "python scripts/run-nonreg.py --tier default --run-all",
