+++ packages/miroir-ai/tsconfig.json
@@ -4,11 +4,9 @@
     "target": "ESNext",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
-    "ignoreDeprecations": "6.0",
     "skipLibCheck": true,
     "outDir": "./dist",
-    "rootDir": "./",
-    "baseUrl": "./"
+    "rootDir": "./"
   },
   "exclude": [
     "./node_modules/**",
