+++ packages/miroir-core/package.json
@@ -13,7 +13,8 @@
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
-    }
+    },
+    "./src/*": "./src/*"
   },
   "files": [
     "LICENSE.md",
