+++ packages/miroir-localcache-redux/package.json
@@ -15,7 +15,7 @@
   },
   "scripts": {
     "watch": "lerna watch --scope='miroir-localcache-redux' --include-dependencies -- lerna run build --scope='miroir-localcache-redux'",
-    "tsbuild": "tsc ./src/index.ts --rootDir . --baseUrl . --target ES2015 --module commonjs --skipLibCheck --strict --moduleResolution node --allowSyntheticDefaultImports --resolveJsonModule --outdir ./dist",
+    "tsbuild": "tsc ./src/index.ts --rootDir . --target ES2015 --module commonjs --skipLibCheck --strict --moduleResolution bundler --allowSyntheticDefaultImports --resolveJsonModule --outdir ./dist",
     "build": "tsup",
     "devBuild": "npm run build",
     "copy-files": "copyfiles -u 1 src/**/*.json src/**/*.css dist/src",
