+++ packages/miroir-react/src/components/Themes/ThemeColorDefaults.ts
@@ -12,7 +12,7 @@
 //
 // ################################################################################################
 
-import type { MiroirThemeFull } from 'miroir-core/src/0_interfaces/1_core/preprocessor-generated/miroirFundamentalType.js';
+import type { MiroirThemeFull } from "miroir-core";
 import { defaultMiroirTheme, MiroirTheme } from "./MiroirTheme.js";
 import { ResolvedTableTheme, TableTheme } from './TableTheme.js';
