+++ packages/miroir-standalone-app-electron/package.json
@@ -11,7 +11,7 @@
   "private": true,
   "scripts": {
     "build-main": "npx tsc src/main.ts --ignoreConfig --outDir dist --target es2020 --module esnext --moduleResolution bundler --allowSyntheticDefaultImports --strict --skipLibCheck",
-    "build-preload": "npx tsc src/preload.ts --ignoreConfig --ignoreDeprecations 6.0 --outDir dist/src --target es2020 --module commonjs --moduleResolution node --allowSyntheticDefaultImports --strict --skipLibCheck",
+    "build-preload": "npx tsc src/preload.ts --ignoreConfig --outDir dist/src --target es2020 --module commonjs --moduleResolution bundler --allowSyntheticDefaultImports --strict --skipLibCheck",
     "build-electron": "npm run build-main && npm run build-preload",
     "build": "npm run build-electron",
     "electron": "npm run build-electron && electron .",
