+++ packages/miroir-standalone-app/src/miroir-fwk/4_view/components/TransformerEditor/TransformerEditor.tsx
@@ -16,8 +16,7 @@ import {
   transformer_extended_apply_wrapper,
   type JzodElement,
   type JzodObject,
-  type JzodUnion,
-  type MiroirModelEnvironment
+  type JzodUnion
 } from 'miroir-core';
 import {
   adminSelfApplication,
@@ -27,14 +26,12 @@ import {
 
 import { Formik, type FormikProps } from 'formik';
 import {
-  type TransformerDefinition,
-  type CoreTransformerForBuildPlusRuntime
-} from "miroir-core/src/0_interfaces/1_core/preprocessor-generated/miroirFundamentalType";
-import { JsonDisplayHelper } from 'miroir-react';
-import { useMiroirContextService } from 'miroir-react';
+  type CoreTransformerForBuildPlusRuntime,
+  type TransformerDefinition
+} from "miroir-core";
+import { JsonDisplayHelper, useMiroirContextService } from 'miroir-react';
 import { packageName } from '../../../../constants';
 import { cleanLevel, lastSubmitButtonClicked } from '../../constants';
-import { useCurrentModelEnvironment } from '../../ReduxHooks';
 import {
   useDeploymentUuidFromApplicationUuid,
   useTransformer
