+++ packages/miroir-store-filesystem/tsconfig.json
@@ -4,14 +4,12 @@
     "outDir": "dist3/out-tsc",
     "target": "ESNext",
     "module": "ESNext",
-    "moduleResolution": "node",
-    "ignoreDeprecations": "6.0",
+    "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "declaration": true,
     "declarationMap": true,
     "declarationDir": "dist3/out-tsc/dts",
     "rootDir": "./",
-    "baseUrl": "./",
   },
   "exclude": ["./node_modules/**","./src/**/*.spec.js", "./src/**/*.spec.jsx", "./src/**/*.spec.ts", "./src/**/*.spec.tsx"],
   "include": ["./src/**/*.json", "./src/**/*.js", "./src/**/*.jsx", "./src/**/*.ts", "./src/**/*.tsx"]
