+++ packages/miroir-store-postgres/package.json
@@ -9,6 +9,14 @@
   "types": "dist/index.d.ts",
   "module": "dist/index.js",
   "type": "module",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    },
+    "./src/*": "./src/*",
+    "./package.json": "./package.json"
+  },
   "scripts": {
     "watch": "lerna watch --scope='miroir-store-postgres' --include-dependencies -- lerna run build --scope='miroir-store-postgres'",
     "build-tsup": "tsup",
