+++ packages/miroir-store-postgres/scripts/postgres-generate-ts-types.ts
@@ -7,8 +7,8 @@ import {
 } from "@miroir-framework/jzod-ts";
 import { sqlQuerySelectSchema } from "../src/1_core/SqlQueryBuilder";
 import { LoggerInterface, MiroirLoggerFactory } from "miroir-core";
-import { cleanLevel } from "miroir-core/src/1_core/constants";
-import { packageName } from "miroir-core/src/constants";
+import { cleanLevel } from "miroir-core/src/1_core/constants.js";
+import { packageName } from "miroir-core/src/constants.js";
 
 let log: LoggerInterface = console as any as LoggerInterface;
 MiroirLoggerFactory.registerLoggerToStart(
